mix of his class edits and more shadin
line 123 is what he did in class with us, that gives saturation and diffuse color correctly, but in black and white because not combined with original color of the hit. When i try to combine with original color, it works for pumpkin but not sheets... not sure why but thinking its something to do with how we multiply color by 255 in line 105. when thats removed the color works well for sheets